TestDashboard is now main dashboard
